requirements.txt
discord.py>=2.6.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
colorlog>=6.7.0
psutil>=5.9.0
asyncio-throttle>=1.0.2
